Update 0500-10-25-tm-user_en.md
The login and sign up processes are described twice: QUICK START GUIDE and Tasking Manager Login. This makes the page unnessecary long, it is already a lot of scrolling to reach the bottom.

So i removed the 2nd description of the login proces. And add a link to the QUICK START GUIDE.

The page is shorter now and a few images that are duplicated are gone. Makes the page imo more usable.