Update 0500-10-12-mapathon_en.md
Toc is partially not working. I am not sure what the root cause is. But i noticed that the items that are not working contain at least one dot.

So i removed the dots and replaced them if applicable with a dash (with escape). Dots in headers are imo not done.

But i couldn't test the outcome: the preview doesn't render the toc.

Please let me know if it worked, thanks.